feat: Added extension to Metadata and used it in MediaInfoCard
The 'ext' field was added to the Metadata data class to store the file extension. This new field is now utilized in the MediaInfoCard composable to display the file extension, replacing the previous direct call to `context.contentResolver.getUriExtension(uri = uri)`. The LocalMediaMetadataDataSource was also updated to populate this new 'ext' field.